+++ drivers/stepper/adi_tmc/CMakeLists.txt
@@ -4,7 +4,6 @@
 zephyr_library()
 zephyr_library_property(ALLOW_EMPTY TRUE)
 
-zephyr_library_sources_ifdef(CONFIG_STEPPER_ADI_TMC_SPI adi_tmc_spi.c)
 zephyr_library_sources_ifdef(CONFIG_STEPPER_ADI_TMC2209 tmc22xx.c)
-zephyr_library_sources_ifdef(CONFIG_STEPPER_ADI_TMC50XX tmc50xx.c)
+zephyr_library_sources_ifdef(CONFIG_ADI_TMC5XXX_STEPPER_DRIVER tmc5xxx_stepper_driver.c)
 zephyr_library_sources_ifdef(CONFIG_STEPPER_ADI_TMC51XX tmc51xx.c)
