+++ dts/bindings/stepper/adi/adi,tmc5xxx_stepper_driver.yaml
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 description: |
-  Analog Devices TMC50XX Stepper Motor Controller
+  Analog Devices TMC50XX Stepper Motor Driver
 
   Example:
 
@@ -54,60 +54,13 @@ description: |
     };
 
 
-compatible: "adi,tmc50xx"
+compatible: "adi,tmc5xxx_stepper_driver"
 
 include:
-  - name: spi-device.yaml
-  - name: adi,trinamic-gconf.yaml
+  - name: stepper.yaml
+  - name: adi,trinamic-stallguard.yaml
     property-allowlist:
-      - poscmp-enable
-      - test-mode
-      - lock-gconf
-
-properties:
-  "#address-cells":
-    default: 1
-    const: 1
-
-  "#size-cells":
-    default: 0
-    const: 0
-
-  clock-frequency:
-    type: int
-    required: true
-    description: |
-      The frequency of the clock signal provided to the TMC50XX.
-      This is used for real world conversion.
-
-      Hint: µstep velocity v[Hz] µsteps / s v[Hz] = v[50xx] * ( fCLK[Hz]/2 / 2^23 )
-            where v[50xx] is the value written to the TMC50XX.
-
-child-binding:
-  include:
-    - name: stepper.yaml
-    - name: base.yaml
-      property-allowlist:
-        - reg
-    - name: adi,trinamic-ramp-generator.yaml
-      property-allowlist:
-        - vstart
-        - a1
-        - v1
-        - amax
-        - vmax
-        - dmax
-        - d1
-        - vstop
-        - tzerowait
-        - vhigh
-        - vcoolthrs
-        - ihold
-        - irun
-        - iholddelay
-    - name: adi,trinamic-stallguard.yaml
-      property-allowlist:
-        - activate-stallguard2
-        - stallguard2-threshold
-        - stallguard-threshold-velocity
-        - stallguard-velocity-check-interval-ms
+      - activate-stallguard2
+      - stallguard2-threshold
+      - stallguard-threshold-velocity
+      - stallguard-velocity-check-interval-ms
