+++ dts/bindings/stepper_control/adi/adi,trinamic-ramp-generator.yaml
