+++ include/zephyr/drivers/stepper/adi_tmc_reg.h
@@ -11,8 +11,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef ZEPHYR_DRIVERS_STEPPER_ADI_TMC_REG_H_
-#define ZEPHYR_DRIVERS_STEPPER_ADI_TMC_REG_H_
+#ifndef ZEPHYR_INCLUDE_DRIVERS_STEPPER_ADI_TMC_REG_H_
+#define ZEPHYR_INCLUDE_DRIVERS_STEPPER_ADI_TMC_REG_H_
 
 #ifdef __cplusplus
 extern "C" {
@@ -21,8 +21,8 @@ extern "C" {
 /** Common Registers for TMC50XX and TMC51XX */
 #if defined(CONFIG_STEPPER_ADI_TMC50XX) || defined(CONFIG_STEPPER_ADI_TMC51XX)
 
-#define TMC5XXX_WRITE_BIT        0x80U
-#define TMC5XXX_ADDRESS_MASK     0x7FU
+#define TMC5XXX_WRITE_BIT    0x80U
+#define TMC5XXX_ADDRESS_MASK 0x7FU
 
 #define TMC5XXX_CLOCK_FREQ_SHIFT 24
 
@@ -34,8 +34,8 @@ extern "C" {
 #define TMC5XXX_RAMPMODE_NEGATIVE_VELOCITY_MODE 2
 #define TMC5XXX_RAMPMODE_HOLD_MODE              3
 
-#define TMC5XXX_SG_MIN_VALUE -64
-#define TMC5XXX_SG_MAX_VALUE 63
+#define TMC5XXX_SG_MIN_VALUE           -64
+#define TMC5XXX_SG_MAX_VALUE           63
 #define TMC5XXX_SW_MODE_SG_STOP_ENABLE BIT(10)
 
 #define TMC5XXX_COOLCONF_SG2_THRESHOLD_VALUE_SHIFT 16
@@ -83,24 +83,24 @@ extern "C" {
 #define TMC50XX_MOTOR_ADDR(m)     (0x20 << (m))
 #define TMC50XX_MOTOR_ADDR_DRV(m) ((m) << 4)
 
-#define TMC50XX_RAMPMODE(motor)   (0x00 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_XACTUAL(motor)    (0x01 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_VACTUAL(motor)    (0x02 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_VSTART(motor)     (0x03 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_A1(motor)         (0x04 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_V1(motor)         (0x05 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_AMAX(motor)       (0x06 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_VMAX(motor)       (0x07 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_DMAX(motor)       (0x08 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_D1(motor)         (0x0A | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_VSTOP(motor)      (0x0B | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_TZEROWAIT(motor)  (0x0C | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_XTARGET(motor)    (0x0D | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_SWMODE(motor)     (0x14 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_RAMPSTAT(motor)   (0x15 | TMC50XX_MOTOR_ADDR(motor))
-#define TMC50XX_CHOPCONF(motor)   (0x6C | TMC50XX_MOTOR_ADDR_DRV(motor))
-#define TMC50XX_COOLCONF(motor)   (0x6D | TMC50XX_MOTOR_ADDR_DRV(motor))
-#define TMC50XX_DRVSTATUS(motor)  (0x6F | TMC50XX_MOTOR_ADDR_DRV(motor))
+#define TMC50XX_RAMPMODE(motor)  (0x00 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_XACTUAL(motor)   (0x01 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_VACTUAL(motor)   (0x02 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_VSTART(motor)    (0x03 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_A1(motor)        (0x04 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_V1(motor)        (0x05 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_AMAX(motor)      (0x06 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_VMAX(motor)      (0x07 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_DMAX(motor)      (0x08 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_D1(motor)        (0x0A | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_VSTOP(motor)     (0x0B | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_TZEROWAIT(motor) (0x0C | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_XTARGET(motor)   (0x0D | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_SWMODE(motor)    (0x14 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_RAMPSTAT(motor)  (0x15 | TMC50XX_MOTOR_ADDR(motor))
+#define TMC50XX_CHOPCONF(motor)  (0x6C | TMC50XX_MOTOR_ADDR_DRV(motor))
+#define TMC50XX_COOLCONF(motor)  (0x6D | TMC50XX_MOTOR_ADDR_DRV(motor))
+#define TMC50XX_DRVSTATUS(motor) (0x6F | TMC50XX_MOTOR_ADDR_DRV(motor))
 
 #endif
 
@@ -150,31 +150,31 @@ extern "C" {
 #define TMC51XX_GCONF_DIAG0_INT_PUSHPULL_SHIFT 12
 #define TMC51XX_GCONF_TEST_MODE_SHIFT          17
 
-#define TMC51XX_IHOLD_IRUN	0x10
-#define TMC51XX_TPOWER_DOWN	0x11
-#define TMC51XX_TSTEP		0x12
-#define TMC51XX_TPWMTHRS	0x13
-#define TMC51XX_TCOOLTHRS	0x14
-#define TMC51XX_THIGH		0x15
+#define TMC51XX_IHOLD_IRUN  0x10
+#define TMC51XX_TPOWER_DOWN 0x11
+#define TMC51XX_TSTEP       0x12
+#define TMC51XX_TPWMTHRS    0x13
+#define TMC51XX_TCOOLTHRS   0x14
+#define TMC51XX_THIGH       0x15
 
-#define TMC51XX_RAMPMODE   	TMC50XX_RAMPMODE(0)
-#define TMC51XX_XACTUAL    	TMC50XX_XACTUAL(0)
-#define TMC51XX_VACTUAL    	TMC50XX_VACTUAL(0)
-#define TMC51XX_VSTART     	TMC50XX_VSTART(0)
-#define TMC51XX_A1         	TMC50XX_A1(0)
-#define TMC51XX_V1         	TMC50XX_V1(0)
-#define TMC51XX_AMAX       	TMC50XX_AMAX(0)
-#define TMC51XX_VMAX       	TMC50XX_VMAX(0)
-#define TMC51XX_DMAX       	TMC50XX_DMAX(0)
-#define TMC51XX_D1         	TMC50XX_D1(0)
-#define TMC51XX_VSTOP      	TMC50XX_VSTOP(0)
-#define TMC51XX_TZEROWAIT  	TMC50XX_TZEROWAIT(0)
-#define TMC51XX_XTARGET    	TMC50XX_XTARGET(0)
-#define TMC51XX_SWMODE     	TMC50XX_SWMODE(0)
-#define TMC51XX_RAMPSTAT   	TMC50XX_RAMPSTAT(0)
-#define TMC51XX_CHOPCONF   	TMC50XX_CHOPCONF(0)
-#define TMC51XX_COOLCONF   	TMC50XX_COOLCONF(0)
-#define TMC51XX_DRVSTATUS  	TMC50XX_DRVSTATUS(0)
+#define TMC51XX_RAMPMODE  TMC50XX_RAMPMODE(0)
+#define TMC51XX_XACTUAL   TMC50XX_XACTUAL(0)
+#define TMC51XX_VACTUAL   TMC50XX_VACTUAL(0)
+#define TMC51XX_VSTART    TMC50XX_VSTART(0)
+#define TMC51XX_A1        TMC50XX_A1(0)
+#define TMC51XX_V1        TMC50XX_V1(0)
+#define TMC51XX_AMAX      TMC50XX_AMAX(0)
+#define TMC51XX_VMAX      TMC50XX_VMAX(0)
+#define TMC51XX_DMAX      TMC50XX_DMAX(0)
+#define TMC51XX_D1        TMC50XX_D1(0)
+#define TMC51XX_VSTOP     TMC50XX_VSTOP(0)
+#define TMC51XX_TZEROWAIT TMC50XX_TZEROWAIT(0)
+#define TMC51XX_XTARGET   TMC50XX_XTARGET(0)
+#define TMC51XX_SWMODE    TMC50XX_SWMODE(0)
+#define TMC51XX_RAMPSTAT  TMC50XX_RAMPSTAT(0)
+#define TMC51XX_CHOPCONF  TMC50XX_CHOPCONF(0)
+#define TMC51XX_COOLCONF  TMC50XX_COOLCONF(0)
+#define TMC51XX_DRVSTATUS TMC50XX_DRVSTATUS(0)
 
 #endif /* CONFIG_STEPPER_ADI_TMC51XX */
 
@@ -186,4 +186,4 @@ extern "C" {
 }
 #endif
 
-#endif /* ZEPHYR_DRIVERS_STEPPER_ADI_TMC_REG_H_ */
+#endif /* ZEPHYR_INCLUDE_DRIVERS_STEPPER_ADI_TMC_REG_H_ */
