+++ samples/drivers/stepper/tmc50xx/prj.conf
@@ -1,2 +1,3 @@
+CONFIG_STEPPER_CONTROL=y
 CONFIG_STEPPER=y
 CONFIG_LOG=y
