+++ samples/drivers/stepper/tmc50xx/src/main.c
@@ -5,21 +5,27 @@
 
 #include <zephyr/device.h>
 #include <zephyr/drivers/gpio.h>
+#include <zephyr/drivers/stepper_control.h>
 #include <zephyr/drivers/stepper.h>
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/stepper/stepper_trinamic.h>
 
-const struct device *stepper = DEVICE_DT_GET(DT_ALIAS(stepper));
+#include <zephyr/logging/log.h>
+LOG_MODULE_REGISTER(tmc50xx_sample, CONFIG_STEPPER_LOG_LEVEL);
 
-int32_t ping_pong_target_position = CONFIG_STEPS_PER_REV * CONFIG_PING_PONG_N_REV *
-				    DT_PROP(DT_ALIAS(stepper), micro_step_res);
+const struct device *stepper = DEVICE_DT_GET(DT_ALIAS(stepper));
+const struct device *stepper_control = DEVICE_DT_GET(DT_ALIAS(stepper_controller));
+
+int32_t ping_pong_target_position =
+	CONFIG_STEPS_PER_REV * CONFIG_PING_PONG_N_REV * DT_PROP(DT_ALIAS(stepper), micro_step_res);
 
 K_SEM_DEFINE(steps_completed_sem, 0, 1);
 
-void stepper_callback(const struct device *dev, const enum stepper_event event, void *user_data)
+void stepper_callback(const struct device *dev, const enum stepper_control_event event,
+		      void *user_data)
 {
 	switch (event) {
-	case STEPPER_EVENT_STEPS_COMPLETED:
+	case STEPPER_CONTROL_EVENT_STEPS_COMPLETED:
 		k_sem_give(&steps_completed_sem);
 		break;
 	default:
@@ -34,21 +40,34 @@ int main(void)
 		printf("Device %s is not ready\n", stepper->name);
 		return -ENODEV;
 	}
-	printf("stepper is %p, name is %s\n", stepper, stepper->name);
+	LOG_DBG("stepper is %p, name is %s\n", stepper, stepper->name);
 
-	stepper_set_event_callback(stepper, stepper_callback, NULL);
+	if (!device_is_ready(stepper_control)) {
+		printf("Device %s is not ready\n", stepper_control->name);
+		return -ENODEV;
+	}
+
+	LOG_DBG("Stepper controller is %p, name is %s\n", stepper_control, stepper_control->name);
+	stepper_control_set_event_callback(stepper_control, stepper_callback, NULL);
+
+	LOG_DBG("Enabling stepper driver %s", stepper->name);
 	stepper_enable(stepper);
-	stepper_set_reference_position(stepper, 0);
-	stepper_move_by(stepper, ping_pong_target_position);
+
+	LOG_DBG("Setting reference position to 0 for %s", stepper_control->name);
+	stepper_control_set_reference_position(stepper_control, 0);
+
+	LOG_DBG("Moving to %d", ping_pong_target_position);
+	stepper_control_move_by(stepper_control, ping_pong_target_position);
 
 	/* Change Max Velocity during runtime */
-	int32_t tmc_velocity = DT_PROP(DT_ALIAS(stepper), vmax) * CONFIG_MAX_VELOCITY_MULTIPLIER;
-	(void)tmc50xx_stepper_set_max_velocity(stepper, tmc_velocity);
+	int32_t tmc_velocity =
+		DT_PROP(DT_ALIAS(stepper_controller), vmax) * CONFIG_MAX_VELOCITY_MULTIPLIER;
+	(void)tmc50xx_stepper_control_set_max_velocity(stepper_control, tmc_velocity);
 
 	for (;;) {
 		if (k_sem_take(&steps_completed_sem, K_FOREVER) == 0) {
 			ping_pong_target_position *= -1;
-			stepper_move_by(stepper, ping_pong_target_position);
+			stepper_control_move_by(stepper_control, ping_pong_target_position);
 		}
 	}
 	return 0;
